% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R
\name{clc}
\alias{clc}
\title{[+] Clear console (Matlab style)}
\usage{
clc()
}
\description{
Clears console. I.e. does the same as pressing \code{CTRL + L}.
}
\seealso{
Other \pkg{spMisc} utilities: \code{\link{bru}},
  \code{\link{clear}}, \code{\link{fCap}}
}

