% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_operators.R
\name{\%++\%}
\alias{\%++\%}
\title{Infix versions of \code{paste0} and \code{paste}}
\source{
Inspired by "Inxix functions" in
\href{http://adv-r.had.co.nz/Functions.html#function-arguments}{Advanced R}
by Hadley Wickham.
}
\usage{
a \%++\% b
}
\arguments{
\item{a, b}{values to be pasted with either \code{paste0(a,b)} or \code{paste}.}
}
\description{
Infix versions of \code{paste0} and \code{paste}.
}
\details{
\code{`\%++\%`} is an infix version of \code{paste0}. \cr
\code{`\%.+.\%`} is an infix version of \code{paste}.
}
\examples{
"a" \%++\% "b"
#> [1] "ab"

"a" \%.+.\% "b"
#> [1] "a b"

}
\seealso{
\code{\link[base]{paste}}, \code{\link[spAddins]{insertPaste0_Addin}}
}

