% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_operators.R
\name{If NULL or length is 0}
\alias{If NULL or length is 0}
\alias{\%if_null_or_len0\%}
\title{[.] Infix operator \%if_null_or_len0\%}
\usage{
a \%if_null_or_len0\% b
}
\arguments{
\item{a}{An expression.}

\item{b}{An alternative.}
}
\value{
If \code{!is.null(a) & length(a) > 0 & nchar(a) > 0} is \code{TRUE}
        returns \code{a}, otherwise returns \code{b}.
}
\description{
The operator is useful as a way of providing a default value
 in case the output of another function is \code{NULL} or its length is
 zero.
}
\examples{
a1 <- "Default value"
a2 <- NULL
a3 <- character()
b  <- "Alternative"

 a1 \%if_null_or_len0\% b
#> [1] "Default value"

 a2 \%if_null_or_len0\% b
#> "Alternative"

 a3 \%if_null_or_len0\% b
#> "Alternative"
}

