% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_functions}
\alias{list_functions}
\title{[+] List all functions in a package}
\usage{
list_functions(Package = "spTMP")
}
\arguments{
\item{Package}{A name of a package. Default \code{Package = "spTMP"}.}
}
\value{
A data frame with function names in a package.
}
\description{
List all functions in a package and return it as a data frame.
}
\examples{

list_functions()
list_functions(Package = "magrittr")

}
\author{
Vilmantas Gegzna
}

