% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexp2df.R
\name{regexp2df}
\alias{regexp2df}
\title{[!+] Capture information to a dataframe by regular expressions}
\usage{
regexp2df(text, pattern, ignore.case = FALSE, perl = TRUE,
  stringsAsFactors = default.stringsAsFactors(), ...)
}
\arguments{
\item{text}{The text to be parsed: a character vector where matches are
sought, or an object which can be coerced by \code{as.character}
to a character vector.}

\item{pattern}{Perl-like regular expression.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{perl}{logical.  Should Perl-compatible regexps be used?}

\item{stringsAsFactors}{logical, passed to \code{\link[base]{as.data.frame}}.}

\item{...}{Other arguments to be passed to \code{\link[base]{gregexpr}}.}
}
\value{
A data frame with extracted information.
}
\description{
Capture information in substrings of \code{text} that match \bold{named} and
\bold{unnamed tokens} of regular expressions and convert the result to a
data frame.
}
\details{
The syntax how to use \bold{named tokens} is: \cr
'...(?<...>...)...' \cr
'expr1(?<Name>expr-to-capture)expr2'

\itemize{
  \item{\bold{Name} - the name of the token.\emph{Note:} that spaces and
  other special symbols,
  inappropriate for variable names, are not allowed and will result in error.}
  \item{\bold{expr-to-capture} - regular expression to be captured as a
  value of a variable.}
  \item{\bold{expr1, expr2} - (optional) expressions, that must match, but
  that are not captured.}
}
\cr

\bold{Unnamed tokens} are simply groups of expressions inside round
brackets "\code{()}". The expressions outside brackets are not captured.
}
\note{
Call to function \code{gregexpr} with parameter \code{perl = TRUE}
is used.
}
\author{
Vilmantas Gegzna
}
\seealso{
More about regular expressions used in R: \link[base]{regex}\cr

         Website handy for creating and testing Perl-like regular expressions
         (library \emph{pcre}, version 1, not 2)
         \url{https://regex101.com/r/dS3iP1/1#pcre} \cr

         Functions \code{\link[base]{gregexpr}},
         operator from package \pkg{magrittr} \code{\link[magrittr]{\%>\%}}.

Other \pkg{spMisc} utilities: \code{\link{bru}},
  \code{\link{clc}}, \code{\link{clear}},
  \code{\link{fCap}}
}

