% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_spMisc-package.R
\docType{package}
\name{spMisc}
\alias{spMisc}
\alias{spMisc-package}
\title{Convenience, wrapper and other kind of miscellaneous functions}
\description{
A set of functions (including convenience and wrapper functions)
designed to be used for variety of different purposes such as
infix operators, Matlab style functions (e.g. that clear objects
 from environments), funcions that find indices #' of certain matrix
elements, functions that sort elements of matrices in a certain way,
functions that modify strings, etc. \cr


[+] - function is well described. \cr
[!] - a description is incomplete and needs revision.\cr\cr

Functions in \code{spMisc} by topic
}
\section{Regular expressions}{


\code{\link{regcapturedmatches}} \cr
\code{\link{regexp2df}} \cr
}

\section{Various}{



\code{\link{bru}} \cr
\code{\link{fCap}} \cr
\code{\link{make_firstCapitals}} \cr
\code{\link{list_functions}} \cr
}
\examples{

\donttest{
\dontrun{

# List all functions in package:

list_functions()

# Plot structure of functions inside the package:
require(sna)
require(mvbutils)

pkgFW <- mvbutils::foodweb(where="package:spMisc", cex=0.7, charlim=60)
sna::gplot(pkgFW$funmat, g = 9,
           jitter = T,
           # mode = "mds",
           label.cex = .6,
           diag=TRUE,
           vertex.cex=1:2,
           displaylabels=TRUE,
           label.bg="gray90")


}}


}
\author{
Vilmantas Gegzna
}

